+++ r-package/tests/testthat/test-download_r5.R
@@ -17,7 +17,6 @@ test_that("download_r5 - expected behavior", {
 
 })
 
-
 # Expected errors
 test_that("download_r5 - expected errors", {